Consolidate cli.rs into main.rs
